ocr_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-vision==3.4.5
pillow==10.1.0
python-multipart==0.0.6

